requirements.txt
langchain==0.3.20
langchain-community==0.3.19
nltk==3.9.1
numpy==2.2.3
rank-bm25==0.2.2
scikit-learn==1.6.1
streamlit==1.43.2
torch==2.6.0
transformers==4.49.0
ipykernel==6.29.5
ipywidgets==8.1.5
pypdf==5.3.1
llama-cpp-python
